13.3.Refactoring Autocomplete JS & CSS{Organizing our Autocomplete into a Component}
Organizing our Autocomplete into a Component

Next, you know the drill, take the algolia-autocomplete.js file and move it into the assets/js/ directory. But I'm not going to make this a new entry point. We could do that, but really, we already have an entry file that's included on this page: admin_article_form.
So really, admin_article_form.js should probably just use the code from algolia-autocomplete.js.
So, move that file into the components/ directory... which is kind of meant for reusable modules. And... well, this isn't really written like a re-usable module yet because it just executes code instead or returning something, like a function. But, we'll work on that later.
Let's also take the algolia-autocomplete.css file and move that all the way up here into assets/css/. And just because we can, I'll make it an SCSS file!
Okay! Back in admin_article_form.js, let's bring in this code: import './components/algolia-autocomplete'
We don't need an import from yet... because that file doesn't actually export anything. For the CSS: import '../css/algolia-autocomplete.scss'.
Back in new.html.twig, the great thing is, we don't need to import this CSS file anymore or any of these script files. This is really how we want our templates to look: a single a call to {{ encore_entry_script_tags() }} and a single call to {{ encore_entry_link_tags() }}.
So if we refresh right now, not surprisingly, it still won't work! And it's our favorite error!

 $ is undefined

from algolia-autocomplete.js. Yes, this is the error I see when I close my eyes at night.