13.4.Refactoring Autocomplete JS & CSS{Using the autocomplete.js Library}
Using the autocomplete.js Library

Let's get to work. Of course, we are referencing $. So, import $ from 'jquery'
We're also using the autocomplete library in here. No problem: import autocomplete from 'autocomplete.js'
Wait... that's not quite right. This autocomplete.js library is a standalone JavaScript library that can be used with anything - jQuery, React, whatever. But... our existing code isn't using the "standalone" version of the library. It's using a jQuery plugin - this .autocomplete() function - that comes with that package:

 23 lines  assets/js/components/algolia-autocomplete.js
$(document).ready(function() {
    $('.js-user-autocomplete').each(function() {
        $(this).autocomplete({hint: false}, [
        ])
    });
});
So, we could refactor our code down here to use the, kind of, official way of using this library - independent of jQuery. But... that's the easy way out! Let's see if we can get this to work as a jQuery plugin.