13.5.Refactoring Autocomplete JS & CSS{Finding and Using the jQuery Plugin}
Finding and Using the jQuery Plugin

I'll hold Command or Control and click into autocomplete.js. Then double-click the directory to zoom us there. The main file is this index.js at the root of the directory. But if you look in dist/, hey! autocomplete.jquery.js! That's what we were including before via the <script> tag!
So instead of importing the main file, let's import autocomplete.js/dist/autocomplete.jquery
And remember, we don't use import from with jQuery plugins... because they don't return anything: they modify the jQuery object.
Ok, I think we're great and I think we're ready. Move over, refresh and... huh:

 jQuery is not defined

Notice it doesn't say "$ is not defined": it says "jQuery is not defined"... and it's coming from autocomplete.jquery.js! It's coming from the third party package!
This... is tricky. Plain and simple, that file is written incorrectly. Yea, it only works if jQuery is a global variable! And in Webpack... it's not! Let's talk more about this and fix it with some black magic, next.