14.1.Auto-Provide jQuery for Mischievous Packages{Poorly-Behaved jQuery Packages}
Everything should be working... but nope! We've got this

  jQuery is not defined
error... but it's not from our code! It's coming from inside of autocomplete.jquery.js - that third party package we installed!

Poorly-Behaved jQuery Packages

This is the second jQuery plugin that we've used. The first was bootstrap... and that worked brilliantly! Look inside app.js:
 26 lines  assets/js/app.js
import $ from 'jquery';
import 'bootstrap'; // adds functions to jQuery
We imported bootstrap and, yea... that was it. Bootstrap is a well-written jQuery plugin, which means that inside, it imports jquery - just like we do - and then modifies it.
But this Algolia autocomplete.js plugin? Yea, it's not so well-written. Instead of detecting that we're inside Webpack and importing jQuery, it just says... jQuery! And expects it to be available as a global variable. This is why jQuery plugins are a special monster: they've been around for so long, that they don't always play nicely in the modern way of doing things.
So... are we stuck? I mean, this 3rd-party package is literally written incorrectly! What can we do?