14.3.Auto-Provide jQuery for Mischievous Packages{Including CSS from the Algolia JS; Making algolia-autocomplete.js a Proper Module}
Take out the import and move it into algolia-autocomplete.js. Make sure to update the path.
That's nice! If we want to use this autocomplete logic somewhere else, we only need to import the JavaScript file: it takes care of importing everything else. The result is the same, but clean.

Making algolia-autocomplete.js a Proper Module

Well, this file still isn't as clean as I want it. We're importing the algolia-autocomplete.js file... but it's not really a "module". It doesn't export some reusable function or class: it just runs code. I really want to start thinking of all of our JavaScript files - except for the entry files themselves - as reusable components.