16.2.Support any Browser with PostCSS & Babel{Enabling PostCSS}
Anyways, forgetting about Webpack and Babel for a minute, in the CSS world, you do not need to add these vendor prefixes by hand. Nope! There's a wonderful library that can do it for you called autoprefixer. You write code correctly - like using box-shadow - tell it which browsers you need to support, and it adds the vendor prefixes for you.

Enabling PostCSS
Because that sounds amazing... let's add it! In webpack.config.js, anywhere, but how about below .enableSassLoader(), add .enablePostCssLoader()
PostCSS is a library that allows you to run things at the "end" of your CSS being processed. And it's the easiest way to integrate autoprefixer.
Next, because we just changed our webpack.config.js file, go restart Encore:
  yarn watch
Hey! This is familiar! Just like when we enabled Sass, this requires us to install a few things.