16.4.Support any Browser with PostCSS & Babel{Enabling PostCSS}
Copy that, go to the root of your project, create the postcss.config.js file and paste.
Ok, hit Control + C and try that again:

yarn watch
Sheesh! One last error. PostCSS is probably the most involved thing to get running. This error isn't as obvious:

  loading PostCSS plugin failed: Cannot find module autoprefixer

We know what that word "module" means! It's trying to find that library. We told PostCSS to use autoprefixer, but that doesn't exist in our project yet