17.2.browserslist: What Browsers do you need to Support?{Configuring Babel}
Configuring Babel

So... what if we change the browserslist config to > 5%?
That's probably still a bit unrealistic: this will only support the most popular browsers and versions: pretty much no old stuff. Stop and re-run Encore:

yarn watch
Then move back over to admin_article_form.js and refresh. I'll do a force refresh to be sure... then search for $autoComplete. And... huh? It's still var? Hmm, that might be right... but const was added in 2015 - it should be fully supported by all modern browsers by now.
It turns out... it is, and we're not seeing the changes due to a small bug in Babel. Behind the scenes, Babel uses some smart caching so that it doesn't need to reparse and recompile every JavaScript file every time Webpack builds. But, at the time of recording, Babel's cache isn't smart enough to know that it needs invalidate itself when the browserslist config changes.
Once you know this, it's no big deal: anytime you change the browserslist config, you need to manually clear Babel's cache. In my terminal, I'll run:

rm -rf node_modules/.cache/babel-loader/
Now restart Encore:

yarn watch
Let's check it out! Refresh and search for . There it is: const . Look also for class ReferenceList. Now that we're only supporting new browsers, that code doesn't need to be rewritten either.
Oh, but there is one type of thing that Babel can't simply rewrite into code that's compatible with olders browsers. When you use a totally new feature of JavaScript - like the fetch() function for AJAX calls, you need to include a polyfill library so that old browsers have this. But... even for this, Babel has a trick up its sleeve. That's next.