19.1.The Single Runtime Chunk
Head back to the homepage and click any of the articles. In an earlier tutorial, we added this heart icon that, when you click it, makes an AJAX request and increases the counter. Well, part of this is faked on the backend, but you get the idea.
To make this more clear, let's add a Bootstrap tooltip: when the user hovers over the heart, we can say something like "Click to like". No problem: open up the template: article/show.html.twig. And I'll remind you that this page has its own entry: article_show.js
 92 lines  templates/article/show.html.twig
{% block javascripts %}
    {{ parent() }}
    {{ encore_entry_script_tags('article_show') }}
{% endblock %}
Go open that: assets/js/article_show.js.
Ok, let's find the anchor tag in the template... there it is... and use multiple lines for sanity. Now add title="Click to Like"
To make this work, all we need to do is copy the js-like-article class, go back to article_show.js and add $('.js-like-article').tooltip(), which is a function added by Bootstrap
Coolio! Let's try it. Refresh and... of course. It doesn't work:

  ...tooltip is not a function

This may or may not surprise you. Think about it: at the bottom of the page, the app.js <script> tags are loaded first. And, if you remember, inside of app.js, we import jquery and then bootstrap, which adds the tooltip() function to jQuery:
 26 lines  assets/js/app.js
import $ from 'jquery';
import 'bootstrap'; // adds functions to jQuery